--- src/nitu/Entity/seller.java
@@ -1,17 +1,0 @@
-package nitu.Entity;
-
-import cn.nukkit.entity.passive.EntityVillager;
-import cn.nukkit.level.format.FullChunk;
-import cn.nukkit.nbt.tag.CompoundTag;
-
-public class seller extends EntityVillager {
-
-
-    public seller(FullChunk chunk, CompoundTag nbt) {
-        super(chunk , nbt);
-    }
-
-
-
-
-}
