--- src/nitu/item/hc.java
@@ -1,14 +1,0 @@
-package nitu.item;
-
-import cn.nukkit.item.Item;
-
-public class hc{
-
-    public Item item() {
-        Item item = new Item(341, 0, 1);
-        String name = "§3回城胶囊";
-        item.setCustomName(name);
-        return item;
-    }
-
-}
