--- src/plugin.yml
@@ -1,10 +1,0 @@
-name: SoilGame
-main: nitu.main
-version: "1.0.0"
-api: ["1.0.0"]
-load: POSTWORLD
-author: 泥土
-description: 泥土大作战
-commands:
- soilgame:
-  description: "/soilgame help 看帮助"
